move MetaInput to its file
