+++ catalog/app/components/MetadataEditor/MetadataEditor.tsx
@@ -2,8 +2,8 @@ import Ajv from 'ajv'
 import brace from 'brace'
 import { JsonEditor as ReactJsonEditor } from 'jsoneditor-react'
 import * as React from 'react'
-import * as Lab from '@material-ui/lab'
 import * as M from '@material-ui/core'
+import * as Lab from '@material-ui/lab'
 
 import JsonEditor from 'components/JsonEditor'
 import { JsonValue } from 'components/JsonEditor/constants'
