+++ catalog/app/containers/Bucket/PackageDialog/PackageCreationForm.tsx
@@ -21,6 +21,7 @@ import DialogSuccess, { DialogSuccessRenderMessageProps } from './DialogSuccess'
 import * as FI from './FilesInput'
 import * as Layout from './Layout'
 import MetaInputErrorHelper from './MetaInputErrorHelper'
+import * as MI from './MetaInput'
 import * as PD from './PackageDialog'
 import { isS3File, S3File } from './S3FilePicker'
 import { FormSkeleton, MetaInputSkeleton } from './Skeleton'
@@ -430,7 +431,7 @@ function PackageCreationForm({
                   ) : (
                     <RF.Field
                       className={classes.meta}
-                      component={PD.MetaInput}
+                      component={MI.MetaInput}
                       name="meta"
                       bucket={bucket}
                       schema={schema}
@@ -438,7 +439,7 @@ function PackageCreationForm({
                       validate={validateMetaInput}
                       validateFields={['meta']}
                       isEqual={R.equals}
-                      initialValue={initial?.manifest?.meta || PD.EMPTY_META_VALUE}
+                      initialValue={initial?.manifest?.meta || MI.EMPTY_META_VALUE}
                       ref={setEditorElement}
                     />
                   )}
